tdsl-netimpl-asio: remove file extension of src/tdsl_netimpl_asio.cxx
Unfortunately, the Arduino Library Index picks up everything listed under the
repository, and that includes the tdslite-net/asio/src/tdsl_netimpl_asio.cxx path.
That was not a problem before because Arduino IDE did not register ".cxx" as a valid
C++ file extension and simply ignored it. But with the Arduino IDE 2.3.2 update that
has changed and all compilations began to fail with a mission "asio header" compilation
error. This patch aims to fix that by simply removing the file's extension and *hoping*
that the Arduino IDE continues to ignore that.

Note that this *hack* had to be put in place because Arduino Library Index lacks the
ability of excluding files while generating a new release artifact from a git tag.

Signed-off-by: Mustafa Kemal GILOR <mustafagilor@gmail.com>